requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cffi==1.17.1
chardet==5.2.0
cssselect2==0.7.0
Django==5.1.4
fonttools==4.55.3
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
psycopg==3.2.3
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.0
python-decouple==3.8
python-dotenv==1.1.0
reportlab==4.2.5
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
weasyprint==63.1
webencodings==0.5.1
whitenoise==6.9.0
zopfli==0.2.3.post1
